Add: domUpdage js. Fix: CSS for page views
